requirements.txt
gradio==4.44.0
matplotlib==3.9.2
numpy==2.1.1
opencv_python==4.9.0.80
Pillow==10.4.0
scikit-image==0.24.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4

